Add new -c/-C options.
-c and -C can be used to specify the keychain to switch to, after
database open.
Extend manual accordingly.
Add new regression tests for the blowfish encryption cipher, scrypt KDF
(if compiled in), and the new -c/-C options.

+++ regress/cmd_export.sh
@@ -18,10 +18,44 @@ case "$(uname -s)" in
 	;;
 esac
 
-rm -f regress/test_export.kcd
 
 PASSWORD=aabbccdd112233
 
+rm -f regress/test_export.kcd
+printf "export -k regress/test_export -e blowfish -m ecb\n${PASSWORD}\n${PASSWORD}\n" |${KC_RUN} -b -k ${KC_DB} -p ${KC_PASSFILE}
+
+if [ ! -r "regress/test_export.kcd" ];then
+	echo "$0 test failed (unreadable export file, blowfish)!"
+	exit 1
+fi
+
+if printf "${PASSWORD}" |${KC_RUN} -b -k regress/test_export.kcd -e blowfish -m ecb;then
+	echo "$0 test ok (export, blowfish)!"
+else
+	echo "$0 test failed (export, blowfish)!"
+	exit 1
+fi
+
+if [ ${SCRYPT} ];then
+	rm -f regress/test_export.kcd
+	printf "export -k regress/test_export -P scrypt\n${PASSWORD}\n${PASSWORD}\n" |${KC_RUN} -b -k ${KC_DB} -p ${KC_PASSFILE}
+
+	if [ ! -r "regress/test_export.kcd" ];then
+		echo "$0 test failed (unreadable export file, scrypt)!"
+		exit 1
+	fi
+
+	if printf "${PASSWORD}" |${KC_RUN} -b -k regress/test_export.kcd -P scrypt;then
+		echo "$0 test ok (export, scrypt)!"
+	else
+		echo "$0 test failed (export, scrypt)!"
+		exit 1
+	fi
+fi
+
+# This test must be the last one that writes test_export.kcd,
+# because the cmd_import.sh test will use this as its input.
+rm -f regress/test_export.kcd
 printf "export -k regress/test_export -P bcrypt -m cfb128\n${PASSWORD}\n${PASSWORD}\n" |${KC_RUN} -b -k ${KC_DB} -p ${KC_PASSFILE}
 
 if [ ! -r "regress/test_export.kcd" ];then
